[DataObject] Add normalize/denormalize method to field definitions
